+++ src/test/java/org/opentmf/db/lock/dialect/DialectOracleHeavyIT.java
@@ -1,0 +1,31 @@
+package org.opentmf.db.lock.dialect;
+
+import javax.sql.DataSource;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.testcontainers.oracle.OracleContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+@Testcontainers
+class DialectOracleHeavyIT extends AbstractDialectIT {
+
+  @Container
+  static final OracleContainer CONTAINER =
+      new OracleContainer(DockerImageName.parse("gvenzl/oracle-free:23-slim-faststart"));
+
+  @Override
+  protected DataSource dataSource() {
+    DriverManagerDataSource ds = new DriverManagerDataSource();
+    ds.setDriverClassName(CONTAINER.getDriverClassName());
+    ds.setUrl(CONTAINER.getJdbcUrl());
+    ds.setUsername(CONTAINER.getUsername());
+    ds.setPassword(CONTAINER.getPassword());
+    return ds;
+  }
+
+  @Override
+  protected Dialect dialect() {
+    return Dialect.ORACLE;
+  }
+}